Removed server and client classes
