+++ thomas/ThomasNet/NetworkManager.cs
@@ -13,41 +13,59 @@ namespace ThomasEngine.Network
     {
         Dictionary<int, NetworkID> networkIDObjects = new Dictionary<int, NetworkID>();
         int iD = -1;
-        Server server = new Server();
-        Client client = new Client();
-        public string IP { get; set; } = "localhost";
+
+      
+        private EventBasedNetListener listener;
+        private NetManager netManager;
+        private NetDataWriter writer;
+        public string IP { get; set; } = "localhost"; 
         public bool Server { get; set; } = false;
         //       public int port { get; set; } = 9050;
-
         public static NetworkManager instance;
         public override void Start()
-        { 
+        {
+            listener = new EventBasedNetListener();
+            netManager = new NetManager(listener);
             instance = this;
-            if(!Server)
-            { 
-                client.SetIp(IP);
-                client.Start();
-                client.ExecuteEvent();
-                //   client.SendDataOverEvent("client send data",DeliveryMethod.ReliableOrdered);
+            writer = new NetDataWriter();
+            if (!Server)
+            {
+                netManager.Start();
+                netManager.Connect(IP /* host ip or name */, 9050 /* port */, "SomeConnectionKey" /* text key or NetDataWriter */);
+
+                //  ExecuteEvent(networkIDObjects);
+                //  SendDataOverEvent("client send data", DeliveryMethod.ReliableOrdered);
+                //  client.SendDataOverEvent(2, DeliveryMethod.ReliableOrdered);
+              //    SendDataOverEvent(0.0f, 5f, 0f, DeliveryMethod.ReliableOrdered);
             }
             else
-            { 
-                server.Start();
-                server.SendDataOverEvent("You are Connected To the server", DeliveryMethod.ReliableOrdered);
-                server.ExecuteEvent();
-            }
+            {
+                netManager.Start(9050 /* port */);
 
+
+                listener.ConnectionRequestEvent += request =>
+                {
+                    if (netManager.PeersCount < 10 /* max connections */)
+                        request.AcceptIfKey("SomeConnectionKey");
+                    else
+                        request.Reject();
+                };
+
+                 SendDataOverEvent("You are Connected To the server", DeliveryMethod.ReliableOrdered);
+
+              //  SendDataOverEvent(0.0f, 2.2f, 0.0f, DeliveryMethod.ReliableOrdered);
+              
+            }
+            ExecuteEvent();
 
         }
         public override void Update()
         {
-            server.Update();
-            client.Update();
+            netManager.PollEvents();
         }
         public override void Destroy()
         {
-            server.Stop();
-            client.Stop();
+            netManager.Stop();
             base.Destroy();
         }
 
@@ -58,6 +76,40 @@ namespace ThomasEngine.Network
             return iD;
         }
 
+        public void ExecuteEvent()
+        {
+            listener.NetworkReceiveEvent += (fromPeer, dataReader, deliveryMethod) =>
+            {
+                 ThomasEngine.Debug.Log(dataReader.GetString());
+                // Console.WriteLine("We got: {0}" , dataReader.GetString(100 /* max length of string */));  dataReader.GetString(100 /* max length of string */)+
+               // networkIDObjects[]
+
+                dataReader.Recycle();
+            };
+        }
+
+        public void SendDataOverEvent(object Data, DeliveryMethod Order)
+        {
+            listener.PeerConnectedEvent += peer =>
+            {
+                //  ThomasEngine.Debug.Log("Server " + peer.EndPoint);// Show peer ip 
+                writer.Put((string)Data);             // Put some string
+                peer.Send(writer, Order);                   // Send with reliability
+                writer.Reset();                             //Resets the writer and emptying its array.
+            };
+        }
+        public void SendDataOverEvent(float Data1, float Data2, float Data3, DeliveryMethod Order)
+        {
+            listener.PeerConnectedEvent += peer =>
+            {
+                //  ThomasEngine.Debug.Log("Server " + peer.EndPoint);// Show peer ip 
+                writer.Put(Data1);
+                writer.Put(Data2);
+                writer.Put(Data3);
+                peer.Send(writer, Order);                   // Send with reliability
+                writer.Reset();                             //Resets the writer and emptying its array.
+            };
+        }
 
     }
 }
